Move : components폴더 통일
