+++ src/app/components/comment/CommentForm.tsx
@@ -1,6 +1,6 @@
 "use client";
 import React, { useState } from "react";
-import { supabase } from "../../utils/supabaseClient";
+import { supabase } from "../../../utils/supabaseClient";
 
 interface CommentFormProps {
   onCommentAdded: () => void;
